Move exceptions back into inner class
