+++ rpki-validator/src/main/java/net/ripe/rpki/validator3/domain/metrics/HttpClientMetricsService.java
@@ -33,8 +33,7 @@ import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Timer;
 import lombok.extern.slf4j.Slf4j;
-import net.ripe.rpki.validator3.util.http.HttpFailureException;
-import net.ripe.rpki.validator3.util.http.HttpStatusException;
+import net.ripe.rpki.validator3.util.HttpStreaming;
 import org.jooq.lambda.tuple.Tuple2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -76,9 +75,9 @@ public class HttpClientMetricsService {
      */
     public static String unwrapExceptionString(Throwable cause) {
         // HttpStatusException is a sub-type of HttpFailureException: check it first.
-        if (cause instanceof HttpStatusException) {
-            return String.valueOf(((HttpStatusException)cause).getCode());
-        } else if (cause instanceof HttpFailureException) {
+        if (cause instanceof HttpStreaming.HttpStatusException) {
+            return String.valueOf(((HttpStreaming.HttpStatusException)cause).getCode());
+        } else if (cause instanceof HttpStreaming.HttpFailureException) {
             final Throwable rootCause = cause.getCause();
             if (rootCause != null) {
                 if (rootCause instanceof EOFException) {
